fix: Let fractional quantities through the API layer
Migration 003 promoted the quantity columns to Numeric(10,4), but the
Pydantic layer still declared int, so the half-carton sales the business
already makes could not be read back or reproduced.

The profit report was failing in production: get_profit_report accumulates
item.quantity into a ProfitReportRow whose quantity was int, so FastAPI's
response validation raised on any date range covering one of the ten
fractional sale items already in the database.

Purchases could not record a fractional quantity at all, and min_stock
could not express a fractional reorder threshold.

+++ app/schemas/purchase.py
@@ -9,12 +9,12 @@ from pydantic import BaseModel, field_validator
 
 class PurchaseItemCreate(BaseModel):
     productId: uuid.UUID
-    quantity: int
+    quantity: Decimal
     unitCost: Decimal
 
     @field_validator("quantity")
     @classmethod
-    def quantity_must_be_positive(cls, v: int) -> int:
+    def quantity_must_be_positive(cls, v: Decimal) -> Decimal:
         if v <= 0:
             raise ValueError("quantity must be greater than 0")
         return v
@@ -48,7 +48,7 @@ class PurchaseUpdate(BaseModel):
 class PurchaseItemResponse(BaseModel):
     id: uuid.UUID
     product_id: uuid.UUID
-    quantity: int
+    quantity: Decimal
     unit_cost: Decimal
     subtotal: Decimal
     # Enriched product info
